move js into modules
