+++ src/scss/partials/vote-slider.scss
@@ -1,0 +1,136 @@
+.vote-bar {
+  position: absolute;
+  padding: 20px 40px 20px 20px;
+  background-color: #fff;
+  max-width: 300px;
+  border-radius: 3px;
+  box-shadow: 0 0 25px rgba(0,0,0,0.2);
+  z-index: 1000;
+}
+.vote__close {
+  position: absolute;
+  transform-origin: center;
+  transform: rotate(45deg);
+  bottom: 2px;
+  right: 0;
+  font-size: 50px;
+  padding: 5px 10px;
+
+}
+.vote__username {
+  width: 100%;
+  margin-bottom: 15px;
+  border-radius: 5px;
+  text-align: center;
+  border: none;
+  border: solid 1px #ccc;
+  padding: 10px;
+}
+.vote__username:focus {
+  outline: none;
+}
+.vote__value {
+  display: inline-block;
+  width: 50px;
+  text-align: center;
+}
+.vote__btn {
+  display: inline-block;
+  border-radius: 50%;
+  height: 30px;
+  width: 30px;
+  text-align: center;
+  line-height: 26px;
+  text-decoration: none;
+}
+.vote__btn svg {
+  transform: translateY(8px);
+}
+
+.vote__slider {
+  width: 140px;
+}
+
+input[type=range] {
+  -webkit-appearance: none;
+  margin: 10px 0;
+}
+input[type=range]:focus {
+  outline: none;
+}
+input[type=range]::-webkit-slider-runnable-track {
+  width: 100%;
+  height: 10px;
+  cursor: pointer;
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+  background: #ccc;
+  border-radius: 25px;
+  border: 0px solid #fafafa;
+}
+input[type=range]::-webkit-slider-thumb {
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+  border: 0px solid #000000;
+  height: 20px;
+  width: 20px;
+  border-radius: 50%;
+  background: #4f79ff;
+  cursor: pointer;
+  -webkit-appearance: none;
+  margin-top: -3.6px;
+}
+input[type=range]:focus::-webkit-slider-runnable-track {
+  background: #ccc;
+}
+input[type=range]::-moz-range-track {
+  width: 100%;
+  height: 10px;
+  cursor: pointer;
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+  background: #ccc;
+  border-radius: 25px;
+  border: 0px solid #fafafa;
+}
+input[type=range]::-moz-range-thumb {
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+  border: 0px solid #000000;
+  height: 20px;
+  width: 20px;
+  border-radius: 50%;
+  background: #4f79ff;
+  cursor: pointer;
+}
+input[type=range]::-ms-track {
+  height: 10px;
+  cursor: pointer;
+  background: transparent;
+  border-color: transparent;
+  border-width: 39px 0;
+  color: transparent;
+}
+input[type=range]::-ms-fill-lower {
+  background: #ccc;
+  border: 0px solid #fafafa;
+  border-radius: 50px;
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+}
+input[type=range]::-ms-fill-upper {
+  background: #ccc;
+  border: 0px solid #fafafa;
+  border-radius: 50px;
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+}
+input[type=range]::-ms-thumb {
+  box-shadow: 0 0 5px rgba(0,0,0,0.1);
+  border: 0px solid #000000;
+  height: 20px;
+  width: 20px;
+  border-radius: 50%;
+  background: #4f79ff;
+  cursor: pointer;
+}
+input[type=range]:focus::-ms-fill-lower {
+  background: #ccc;
+}
+input[type=range]:focus::-ms-fill-upper {
+  background: #ccc;
+}
